[frontend] Item variable for empty value
