+++ opencti-platform/opencti-front/src/components/ItemDueDate.tsx
@@ -1,40 +1,24 @@
 import React from 'react';
-import Chip from '@mui/material/Chip';
-import makeStyles from '@mui/styles/makeStyles';
+import Tag from '@common/tag/Tag';
 import { useFormatter } from './i18n';
-import { chipInListBasicStyle } from '../utils/chipStyle';
-
-// Deprecated - https://mui.com/system/styles/basics/
-// Do not use it for new code.
-const useStyles = makeStyles(() => ({
-  chip: {
-    borderRadius: 4,
-  },
-  chipInList: {
-    ...chipInListBasicStyle,
-    width: 120,
-  },
-}));
+import { EMPTY_VALUE } from '../utils/String';
+import { useTheme } from '@mui/material/styles';
 
 const ItemDueDate = ({ due_date, variant }: { due_date: string | null; variant: string }) => {
   const { fld, fldt } = useFormatter();
-  const classes = useStyles();
+  const theme = useTheme();
   const isoDate = new Date().toISOString();
-  const style = variant === 'inList' ? classes.chipInList : classes.chip;
   const label = variant === 'inList' ? fld(due_date) : fldt(due_date);
   if (due_date) {
     return (
-      <Chip
-        classes={{ root: style }}
+      <Tag
         variant="outlined"
         label={label}
-        color={due_date < isoDate ? 'error' : 'info'}
+        color={due_date < isoDate ? theme.palette.error.main : theme.palette.primary.main}
       />
     );
   }
-  return (
-    <>-</>
-  );
+  return <>{EMPTY_VALUE}</>;
 };
 
 export default ItemDueDate;
