+++ opencti-platform/opencti-front/src/private/components/cases/case_rfts/CaseRft.tsx
@@ -16,7 +16,7 @@ import CaseRftDetails from './CaseRftDetails';
 import { useFormatter } from '../../../../components/i18n';
 import { usePaginationLocalStorage } from '../../../../utils/hooks/useLocalStorage';
 import ListLines from '../../../../components/list_lines/ListLines';
-import { tasksDataColumns } from '../tasks/TasksLine';
+import { tasksDataColumns } from '../tasks/tasksDataColumns';
 import { CaseTasksLineDummy } from '../tasks/CaseTasksLine';
 import { isFilterGroupNotEmpty, useRemoveIdAndIncorrectKeysFromFilterGroupObject } from '../../../../utils/filters/filtersUtils';
 import { FilterGroup } from '../../../../utils/filters/filtersHelpers-types';
